Move TimerContext to sys module and add new one single thread to snapshot for derby.
