+++ persistence/src/main/java/com/alibaba/nacos/persistence/utils/PersistenceExecutor.java
@@ -20,6 +20,7 @@ import com.alibaba.nacos.common.executor.ExecutorFactory;
 import com.alibaba.nacos.common.executor.NameThreadFactory;
 
 import java.util.concurrent.Executor;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
@@ -36,7 +37,11 @@ public class PersistenceExecutor {
     
     private static final Executor DUMP_EXECUTOR = ExecutorFactory.Managed
             .newSingleExecutorService(PersistenceExecutor.class.getCanonicalName(),
-                    new NameThreadFactory("com.alibaba.nacos.config.embedded.dump"));
+                    new NameThreadFactory("com.alibaba.nacos.persistence.embedded.dump"));
+    
+    private static final ExecutorService EMBEDDED_SNAPSHOT_EXECUTOR = ExecutorFactory.Managed
+            .newSingleExecutorService(PersistenceExecutor.class.getCanonicalName(),
+                    new NameThreadFactory("com.alibaba.nacos.persistence.embedded.snapshot"));
     
     public static void scheduleTask(Runnable command, long initialDelay, long delay, TimeUnit unit) {
         TIMER_EXECUTOR.scheduleWithFixedDelay(command, initialDelay, delay, unit);
@@ -44,5 +49,9 @@ public class PersistenceExecutor {
     
     public static void executeEmbeddedDump(Runnable runnable) {
         DUMP_EXECUTOR.execute(runnable);
+    }
+    
+    public static void executeSnapshot(Runnable runnable) {
+        EMBEDDED_SNAPSHOT_EXECUTOR.execute(runnable);
     }
 }
